+++ lib/mdex/sigil.ex
@@ -37,9 +37,9 @@ defmodule MDEx.Sigil do
     * `XML` - converts Markdown or `MDEx.Document` to XML, equivalent to calling `MDEx.to_xml!/2`
     * `MD` - converts `MDEx.Document` to Markdown, equivalent to calling `MDEx.to_markdown!/2`
 
-  Without a modifier it converts Markdown to `MDEx.Document` (the default output of `~m` and `~M`), equivalent to calling `MDEx.parse_document!/2`.
+  Defaults to generating a `MDEx.Document` when no modifier is provided.
 
-  Note that you should `import MDEx.Sigil` to use the `~m` or `~M` sigil.
+  Note that you should `import MDEx.Sigil` to use the `~MD` sigil.
 
   ## Options
 
@@ -54,48 +54,79 @@ defmodule MDEx.Sigil do
   """
 
   @doc """
-  The `~M` sigil converts to `MDEx.Document`, CommonMark, HTML, JSON or XML without interpolation.
+  The `~MD` sigil converts a Markdown string or a `%MDEx.Document{}` struct to either one of these formats: `MDEx.Document`, Markdown (CommonMark), HTML, JSON or XML.
+
+  ## Assigns
+
+  You can define a variable `assigns` in the context of the sigil to pass values to the Markdown string, for example:
+
+      assigns = %{lang: ":elixir"}
+      iex> ~MD|`lang = <%= @lang %>`|
+      %MDEx.Document{nodes: [%MDEx.Paragraph{nodes: [%MDEx.Code{num_backticks: 1, literal: "lang = :elixir"}]}]}
 
   ## Examples
 
   ### Markdown to `MDEx.Document`
 
+  No modifier defaults to generating a `MDEx.Document`:
+
   ```elixir
-  iex> ~M[`lang = :elixir`]
+  iex> ~MD[`lang = :elixir`]
   %MDEx.Document{nodes: [%MDEx.Paragraph{nodes: [%MDEx.Code{num_backticks: 1, literal: "lang = :elixir"}]}]}
   ```
 
   ### Markdown to HTML
 
   ```elixir
-  iex> ~M[`lang = :elixir`]HTML
+  iex> ~MD[`lang = :elixir`]HTML
   "<p><code>lang = :elixir</code></p>\\n"
   ```
 
   ### Markdown to JSON
 
   ```elixir
-  iex> ~M[`lang = :elixir`]JSON
+  iex> ~MD[`lang = :elixir`]JSON
   "{\"nodes\":[{\"nodes\":[{\"literal\":\"lang = :elixir\",\"num_backticks\":1,\"node_type\":\"MDEx.Code\"}],\"node_type\":\"MDEx.Paragraph\"}],\"node_type\":\"MDEx.Document\"}"
   ```
 
   ### Markdown to XML
 
   ```elixir
-  iex> ~M[`lang = :elixir`]XML
+  iex> ~MD[`lang = :elixir`]XML
   "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<!DOCTYPE document SYSTEM \"CommonMark.dtd\">\n<document xmlns=\"http://commonmark.org/xml/1.0\">\n  <paragraph>\n    <code xml:space=\"preserve\">lang = :elixir</code>\n  </paragraph>\n</document>\n"
   ```
 
   ### `MDEx.Document` to Markdown
 
   ```elixir
-  iex> ~M|%MDEx.Document{nodes: [%MDEx.Paragraph{nodes: [%MDEx.Code{num_backticks: 1, literal: "lang = :elixir"}]}]}|MD
+  iex> ~MD|%MDEx.Document{nodes: [%MDEx.Paragraph{nodes: [%MDEx.Code{num_backticks: 1, literal: "lang = :elixir"}]}]}|MD
   "`lang = :elixir`"
   ```
 
-  Other modifiers also convert from a document to the desired format.
-
   """
+  defmacro sigil_MD({:<<>>, _meta, [expr]}, modifiers) do
+    expr = expr(expr, __CALLER__)
+
+    quote generated: true do
+      cond do
+        unquote(modifiers) == [] -> MDEx.parse_document!(unquote(expr), unquote(@opts))
+        unquote(modifiers) == ~c"HTML" -> MDEx.to_html!(unquote(expr), unquote(@opts))
+        unquote(modifiers) == ~c"JSON" -> MDEx.to_json!(unquote(expr), unquote(@opts))
+        unquote(modifiers) == ~c"XML" -> MDEx.to_xml!(unquote(expr), unquote(@opts))
+        unquote(modifiers) == ~c"MD" -> MDEx.to_markdown!(unquote(expr), unquote(@opts))
+      end
+    end
+  end
+
+  defp expr(expr, env) do
+    with {:ok, {:%, _, _} = quoted} <- Code.string_to_quoted(expr) do
+      expand_alias(quoted, env)
+    else
+      _ -> EEx.compile_string(expr)
+    end
+  end
+
+  @deprecated "Use the ~MD sigil instead"
   defmacro sigil_M({:<<>>, _meta, [expr]}, modifiers) do
     expr = Macro.unescape_string(expr)
     doc = to_doc(expr, __CALLER__)
@@ -131,26 +162,7 @@ defmodule MDEx.Sigil do
     Macro.escape(doc)
   end
 
-  @doc """
-  The `~m` sigil converts to `MDEx.Document`, CommonMark, HTML, JSON or XML with interpolation.
-
-  ## Examples
-
-  ### Markdown to `MDEx.Document`
-
-      iex> lang = :elixir
-      iex> ~m[`lang = \#{inspect(lang)}`]
-      %MDEx.Document{nodes: [%MDEx.Paragraph{nodes: [%MDEx.Code{num_backticks: 1, literal: "lang = :elixir"}]}]}
-
-  ### `MDEx.Document` to Markdown
-
-      iex> lang = :elixir
-      iex> ~m|%MDEx.Document{nodes: [%MDEx.Paragraph{nodes: [%MDEx.Code{num_backticks: 1, literal: "lang = \#{inspect(lang)}"}]}]}|
-      "`lang = :elixir`"
-
-  Other modifiers also accept interpolation.
-
-  """
+  @deprecated "Use the ~MD sigil instead"
   defmacro sigil_m({:<<>>, _, [binary]}, modifiers) when is_binary(binary) do
     binary = Macro.unescape_string(binary)
     doc = to_doc(binary, __CALLER__)